- Fixed build - Moved checksum to Mosa.Kernel.x86
