+++ Source/Mosa.ClassLib/Mosa.ClassLib.csproj
@@ -61,7 +61,6 @@
     <Compile Include="BinaryFormat.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="FlechtersChecksum.cs" />
     <Compile Include="Math.cs">
       <SubType>Code</SubType>
     </Compile>
